Add additional system tests
